detail

+++ xinference/web/ui/src/scenes/launch_model/modelCard.js
@@ -282,16 +282,21 @@ const ModelCard = ({
 
   const handeCustomDelete = (e) => {
     e.stopPropagation()
-    fetcher(url + `/v1/model_registrations/LLM/${modelData.model_name}`, {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-    })
-      .then(() => {
-        setCustomDeleted(true)
+    const subType = sessionStorage.getItem('subType').split('/')
+    if(subType) {
+
+      subType[3]
+      fetcher(url + `/v1/model_registrations/${subType[3] === 'llm'? 'LLM' : subType[3]}/${modelData.model_name}`, {
+        method: 'DELETE',
+        headers: {
+          'Content-Type': 'application/json',
+        },
       })
-      .catch(console.error)
+        .then(() => {
+          setCustomDeleted(true)
+        })
+        .catch(console.error)
+    }
   }
 
   const judgeArr = (arr, keysArr) => {
@@ -373,7 +378,7 @@ const ModelCard = ({
             flexWrap="wrap"
             sx={{ marginLeft: 1 }}
           >
-            {(() => {
+            {modelData.model_lang && (() => {
               return modelData.model_lang.map((v) => {
                 return (
                   <Chip key={v} label={v} variant="outlined" size="small" />
@@ -381,7 +386,7 @@ const ModelCard = ({
               })
             })()}
             {(() => {
-              if (modelData.model_specs.some((spec) => isCached(spec))) {
+              if (modelData.model_specs && modelData.model_specs.some((spec) => isCached(spec))) {
                 return <Chip label="Cached" variant="outlined" size="small" />
               }
             })()}
@@ -405,14 +410,14 @@ const ModelCard = ({
               <small style={styles.smallText}>context length</small>
             </div>
             {(() => {
-              if (modelData.model_ability.includes('chat')) {
+              if (modelData.model_ability && modelData.model_ability.includes('chat')) {
                 return (
                   <div style={styles.iconItem}>
                     <ChatOutlined style={styles.muiIcon} />
                     <small style={styles.smallText}>chat model</small>
                   </div>
                 )
-              } else if (modelData.model_ability.includes('generate')) {
+              } else if (modelData.model_ability && modelData.model_ability.includes('generate')) {
                 return (
                   <div style={styles.iconItem}>
                     <EditNoteOutlined style={styles.muiIcon} />
